Added error checking in many places. + cleanup.
Also added GamepadConnected & GamepadDisconnected events.
And some other minor things.

+++ include/TML/Graphics/Drawable/Line.h
@@ -7,7 +7,7 @@ namespace tml
     {
     public:
         Line();
-        Line(const Vector2f& a, const Vector2f& b, uint32_t thickness = 1, const Color& color = Color::White);
+        Line(const Vector2f& a, const Vector2f& b, float thickness = 1, const Color& color = Color::White);
 
         void SetPointA(const Vector2f& point) noexcept;
         void SetPointB(const Vector2f& point) noexcept;
